Feat: analytics_screen에 차트 삽입
추가적으로 'analytics-screen' 에서 'analytics_screen'으로 파일명을 변경하였습니다.
